+++ dd/mdd.py
@@ -24,7 +24,6 @@ import sys
 
 from dd import bdd as _bdd
 from dd.bdd import to_nx
-from dd._compat import items
 # inline:
 # import networkx
 # import pydot
@@ -33,11 +32,6 @@ from dd._compat import items
 logger = logging.getLogger(__name__)
 TABMODULE = 'dd.mdd_parsetab'
 PLY_LOG = 'dd.mdd.ply'
-# for python 3
-try:
-    xrange(0)
-except NameError:
-    xrange = range
 
 
 class MDD(object):
@@ -82,11 +76,7 @@ class MDD(object):
         self._level_to_var = None
         self._parser = None
         self._free = set()
-        try:
-            # Python 2, for xrange
-            self.max_nodes = sys.maxint
-        except AttributeError:
-            self.max_nodes = sys.maxsize
+        self.max_nodes = sys.maxsize
 
     def __len__(self):
         """Return number of BDD nodes."""
@@ -134,7 +124,7 @@ class MDD(object):
         if self._level_to_var is None:
             self._level_to_var = {
                 d['level']: var
-                for var, d in items(self.vars)}
+                for var, d in self.vars.items()}
         return self._level_to_var[i]
 
     def level_of_var(self, var):
@@ -255,7 +245,7 @@ class MDD(object):
             cond[x].add(j)
         # format
         cond_str = dict()
-        for k, v in items(cond):
+        for k, v in cond.items():
             if len(v) == 1:
                 (j,) = v
                 cond_str[k] = '= {j}'.format(j=j)
@@ -356,15 +346,15 @@ def bdd_to_mdd(bdd, dvars):
     #
     # map from bits to integers
     bit_to_var = dict()
-    for var, d in items(dvars):
+    for var, d in dvars.items():
         bits = d['bitnames']
         b = {bit: var for bit in bits}
         bit_to_var.update(b)
     # find target bit order
     order = list()  # target
-    levels = {d['level']: var for var, d in items(dvars)}
+    levels = {d['level']: var for var, d in dvars.items()}
     m = len(levels)
-    for j in xrange(m):
+    for j in range(m):
         var = levels[j]
         bits = dvars[var]['bitnames']
         order.extend(bits)
@@ -376,7 +366,7 @@ def bdd_to_mdd(bdd, dvars):
     mdd = MDD(dvars)
     # zones of bits per integer var
     zones = dict()
-    for var, d in items(dvars):
+    for var, d in dvars.items():
         bits = d['bitnames']
         lsb = bits[0]
         msb = bits[-1]
@@ -385,7 +375,7 @@ def bdd_to_mdd(bdd, dvars):
         zones[var] = (min_level, max_level)
     # reverse edges
     pred = {u: set() for u in bdd}
-    for u, (_, v, w) in items(bdd._succ):
+    for u, (_, v, w) in bdd._succ.items():
         assert u > 0, u
         # terminal ?
         if u == 1:
@@ -395,7 +385,7 @@ def bdd_to_mdd(bdd, dvars):
         pred[abs(w)].add(u)
     # find BDD nodes mentioned from above
     rm = set()
-    for u, p in items(pred):
+    for u, p in pred.items():
         rc = bdd.ref(u)
         k = len(p)  # number of predecessors
         # has external refs ?
@@ -412,7 +402,7 @@ def bdd_to_mdd(bdd, dvars):
             continue
         # referenced only from inside zone
         rm.add(u)
-    pred = {u: p for u, p in items(pred) if u not in rm}
+    pred = {u: p for u, p in pred.items() if u not in rm}
     # build layer by layer
     # TODO: use bins, instad of iterating through all nodes
     bdd.assert_consistent()
@@ -446,7 +436,7 @@ def bdd_to_mdd(bdd, dvars):
 
 def _enumerate_integer(bits):
     n = len(bits)
-    for i in xrange(2**n):
+    for i in range(2**n):
         values = list(reversed(bin(i).lstrip('-0b').zfill(n)))
         d = {bit: int(v) for bit, v in zip(bits, values)}
         for bit in bits[len(values):]:
@@ -481,7 +471,7 @@ def to_pydot(mdd):
     skeleton = list()
     subgraphs = dict()
     n = len(mdd.vars) + 1
-    for i in xrange(n):
+    for i in range(n):
         h = pydot.Subgraph('', rank='same')
         g.add_subgraph(h)
         subgraphs[i] = h
@@ -496,7 +486,7 @@ def to_pydot(mdd):
         e = pydot.Edge(str(u), str(v), style='invis')
         g.add_edge(e)
     # add nodes
-    for u, t in items(mdd._succ):
+    for u, t in mdd._succ.items():
         assert u > 0, u
         i = t[0]
         nodes = t[1:]